+++ src/feature.rs
@@ -56,6 +56,7 @@ where
     PercentDifferenceMagnitudePercentile,
     Periodogram(Periodogram<T, Self>),
     _PeriodogramPeaks(PeriodogramPeaks),
+    QnScale,
     ReducedChi2,
     Roms,
     Skew,